+++ app/src/androidTest/kotlin/cz/covid19cz/erouska/screenObject/HomeScreen.kt
@@ -2,6 +2,9 @@ package cz.covid19cz.erouska.screenObject
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.matcher.ViewMatchers.*
 import cz.covid19cz.erouska.R
+import cz.covid19cz.erouska.helpers.checkDisplayed
+import cz.covid19cz.erouska.helpers.checkMatchesString
+import cz.covid19cz.erouska.helpers.click
 import org.awaitility.Awaitility.await
 import org.hamcrest.CoreMatchers.endsWith
 import java.util.concurrent.TimeUnit
