+++ app/src/androidTest/kotlin/cz/covid19cz/erouska/screenObject/WelcomeScreen.kt
@@ -1,10 +1,11 @@
 package cz.covid19cz.erouska.screenObject
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import cz.covid19cz.erouska.R
+import cz.covid19cz.erouska.helpers.checkMatchesContainsString
+import cz.covid19cz.erouska.helpers.click
 
 object WelcomeScreen {
     const val PART_OF_HELP_TITLE = "Proč je eRouška potřeba?"
